+++ components/eventForm/SectionHeader.tsx
@@ -2,7 +2,7 @@ import React, { ReactNode } from "react"
 import { StyleSheet, Text, TouchableOpacity, View } from "react-native"
 import { FontScaleFactors } from "../../lib/FontScale"
 import { useEventFormContext } from "./EventForm"
-import { MaterialIcon } from "@components/common/MaterialIcon"
+import { MaterialIcon } from "@components/common/Icons"
 
 export type EventFormSectionHeaderProps = {
   title: string
